refactor: centralize tool detail behavior
