+++ src/server_msg.rs
@@ -6,18 +6,11 @@
 //! compatibility tests and as a reference while old event-shaped behavior is
 //! ported to native ACP state.
 
-use std::collections::{HashMap, HashSet};
-use std::path::{Path, PathBuf};
-
 use crate::app::*;
 use crate::protocol::*;
 use crate::tool_detail;
-use crate::ui::{
-    ELLIPSIS, OUTCOME_BULLET, build_diff_lines, build_sectioned_diff_lines, build_shell_lines,
-    build_write_lines,
-};
-
-const DEFAULT_READ_TOOL_LIMIT: u64 = 2000;
+use crate::ui::OUTCOME_BULLET;
+use std::collections::{HashMap, HashSet};
 
 #[derive(Debug, Clone)]
 pub(crate) struct PendingElicitationSnapshot {
@@ -2243,567 +2236,31 @@ pub(crate) fn backfill_elicitation_outcomes(messages: &mut [ChatEntry], result_s
     }
 }
 
-fn truncate_summary(s: &str, max_bytes: usize) -> String {
-    if s.len() <= max_bytes {
-        return s.to_string();
-    }
-
-    let end = s
-        .char_indices()
-        .map(|(i, _)| i)
-        .take_while(|&i| i <= max_bytes)
-        .last()
-        .unwrap_or(0);
-    format!("{}{ELLIPSIS}", &s[..end])
-}
-
-fn short_path_display(path: &str) -> String {
-    let mut count = 0;
-    for (i, c) in path.char_indices().rev() {
-        if c == '/' {
-            count += 1;
-            if count == 2 {
-                return path[i + 1..].to_string();
-            }
-        }
-    }
-    path.to_string()
-}
-
-fn scalar_string(obj: &serde_json::Value, key: &str) -> Option<String> {
-    obj.get(key).and_then(|v| v.as_str()).map(str::to_string)
-}
-
-fn shell_command_display(obj: &serde_json::Value) -> String {
-    let command = scalar_string(obj, "command").unwrap_or_default();
-    let Some(args) = obj.get("args").and_then(|v| v.as_array()) else {
-        return command;
-    };
-
-    let mut parts = Vec::with_capacity(args.len() + 1);
-    parts.push(shell_quote_arg(&command));
-    parts.extend(
-        args.iter()
-            .filter_map(|arg| arg.as_str())
-            .map(shell_quote_arg),
-    );
-
-    if parts.len() == 1 {
-        return command;
-    }
-
-    parts
-        .iter()
-        .enumerate()
-        .map(|(idx, part)| {
-            let continuation = if idx + 1 == parts.len() { "" } else { " \\" };
-            format!("{part}{continuation}")
-        })
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-fn shell_quote_arg(arg: &str) -> String {
-    if arg.is_empty() {
-        return "''".to_string();
-    }
-    if arg.chars().all(|c| {
-        c.is_ascii_alphanumeric()
-            || matches!(c, '_' | '-' | '.' | '/' | ':' | '=' | '+' | ',' | '@' | '%')
-    }) {
-        return arg.to_string();
-    }
-    format!("'{}'", arg.replace('\'', "'\\''"))
-}
-
-fn resolve_preview_path(path: &str, root: Option<&str>, cwd: Option<&str>) -> PathBuf {
-    let path_buf = PathBuf::from(path);
-    if path_buf.is_absolute() {
-        return path_buf;
-    }
-
-    let base = root
-        .filter(|root| !root.trim().is_empty())
-        .map(PathBuf::from)
-        .unwrap_or_else(|| cwd.map(PathBuf::from).unwrap_or_else(|| PathBuf::from(".")));
-    let base = if base.is_absolute() {
-        base
-    } else if let Some(cwd) = cwd {
-        Path::new(cwd).join(base)
-    } else {
-        base
-    };
-
-    base.join(path_buf)
-}
-
-fn replace_symbol_title(replacements: &[serde_json::Value]) -> String {
-    if replacements.len() == 1 {
-        let replacement = &replacements[0];
-        let path = replacement
-            .get("path")
-            .and_then(|v| v.as_str())
-            .unwrap_or("");
-        let symbol = replacement
-            .get("symbol")
-            .and_then(|v| v.as_str())
-            .unwrap_or("");
-        if path.is_empty() {
-            return symbol.to_string();
-        }
-        if symbol.is_empty() {
-            return short_path_display(path);
-        }
-        return format!("{} {}", short_path_display(path), symbol);
-    }
-
-    let files = replacements
-        .iter()
-        .filter_map(|r| r.get("path").and_then(|v| v.as_str()))
-        .collect::<HashSet<_>>()
-        .len();
-    format!("{} symbols in {} files", replacements.len(), files)
-}
-
-fn build_replace_symbol_sections(
-    replacements: &[serde_json::Value],
-    root: Option<&str>,
-    cwd: Option<&str>,
-) -> Vec<DiffPreviewSection> {
-    replacements
-        .iter()
-        .filter_map(|replacement| {
-            let path = replacement.get("path").and_then(|v| v.as_str())?;
-            let symbol = replacement.get("symbol").and_then(|v| v.as_str())?;
-            let new_text = replacement.get("newText").and_then(|v| v.as_str())?;
-            let kind = replacement.get("kind").and_then(|v| v.as_str());
-            let occurrence = replacement
-                .get("occurrence")
-                .and_then(|v| v.as_u64())
-                .unwrap_or(0) as usize;
-            let resolved_path = resolve_preview_path(path, root, cwd);
-            let (old, start_line) = extract_symbol_body(&resolved_path, symbol, kind, occurrence)
-                .unwrap_or_else(|| (String::new(), 1));
-            let header = if replacements.len() == 1 {
-                symbol.to_string()
-            } else {
-                format!("{} {}", short_path_display(path), symbol)
-            };
-            let header = if let Some(kind) = kind.filter(|kind| !kind.is_empty() && *kind != "any")
-            {
-                format!("{header} ({kind})")
-            } else {
-                header
-            };
-            Some(DiffPreviewSection {
-                header,
-                old,
-                new: new_text.to_string(),
-                start_line: Some(start_line),
-            })
-        })
-        .collect()
-}
-
-fn extract_symbol_body(
-    path: &Path,
-    symbol: &str,
-    kind: Option<&str>,
-    occurrence: usize,
-) -> Option<(String, usize)> {
-    let content = std::fs::read_to_string(path).ok()?;
-    let lines: Vec<&str> = content.lines().collect();
-    let leaf = symbol
-        .rsplit("::")
-        .next()
-        .unwrap_or(symbol)
-        .rsplit('.')
-        .next()
-        .unwrap_or(symbol);
-
-    let starts = lines
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, line)| symbol_line_matches(line.trim_start(), leaf, kind).then_some(idx))
-        .collect::<Vec<_>>();
-    let start = *starts.get(occurrence)?;
-    let end = symbol_block_end(&lines, start);
-    Some((lines[start..=end].join("\n"), start + 1))
-}
-
-fn symbol_line_matches(line: &str, name: &str, kind: Option<&str>) -> bool {
-    let matches_function = || {
-        line.contains(&format!("fn {name}"))
-            || line.contains(&format!("function {name}"))
-            || line.starts_with(&format!("def {name}"))
-            || line.starts_with(&format!("async def {name}"))
-            || line.starts_with(&format!("{name}("))
-            || line.contains(&format!(" {name}("))
-    };
-    let matches_struct = || line.contains(&format!("struct {name}"));
-    let matches_enum = || line.contains(&format!("enum {name}"));
-    let matches_trait = || line.contains(&format!("trait {name}"));
-    let matches_class = || line.contains(&format!("class {name}"));
-    let matches_type = || line.contains(&format!("type {name}"));
-    let matches_const = || line.contains(&format!("const {name}"));
-
-    match kind.unwrap_or("any") {
-        "function" | "method" | "test" => matches_function(),
-        "struct" => matches_struct(),
-        "enum" => matches_enum(),
-        "trait" => matches_trait(),
-        "class" => matches_class(),
-        "type" => matches_type(),
-        "const" => matches_const(),
-        _ => {
-            matches_function()
-                || matches_struct()
-                || matches_enum()
-                || matches_trait()
-                || matches_class()
-                || matches_type()
-                || matches_const()
-        }
-    }
-}
-
-fn symbol_block_end(lines: &[&str], start: usize) -> usize {
-    let mut seen_brace = false;
-    let mut balance = 0isize;
-    for (idx, line) in lines.iter().enumerate().skip(start) {
-        for ch in line.chars() {
-            match ch {
-                '{' => {
-                    seen_brace = true;
-                    balance += 1;
-                }
-                '}' if seen_brace => {
-                    balance -= 1;
-                    if balance <= 0 {
-                        return idx;
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    if seen_brace {
-        return lines.len().saturating_sub(1);
-    }
-
-    let start_indent = line_indent(lines[start]);
-    let mut end = start;
-    for (idx, line) in lines.iter().enumerate().skip(start + 1) {
-        if line.trim().is_empty() {
-            end = idx;
-            continue;
-        }
-        if line_indent(line) <= start_indent {
-            break;
-        }
-        end = idx;
-    }
-    end
-}
-
-fn line_indent(line: &str) -> usize {
-    line.chars().take_while(|c| c.is_whitespace()).count()
-}
-
-fn shell_output_tail_from_result(
-    parsed_result: Option<&serde_json::Value>,
-    raw_result: &str,
-) -> Option<ShellOutputTail> {
-    let (stdout, stderr) = parsed_result
-        .and_then(|obj| {
-            Some((
-                obj.get("stdout")?.as_str().unwrap_or_default().to_string(),
-                obj.get("stderr")
-                    .and_then(|v| v.as_str())
+fn content_to_string(value: &serde_json::Value) -> String {
+    match value {
+        serde_json::Value::String(text) => text.clone(),
+        serde_json::Value::Object(object) => {
+            if object.get("type").and_then(serde_json::Value::as_str) == Some("text") {
+                object
+                    .get("text")
+                    .and_then(serde_json::Value::as_str)
+                    .map(str::to_string)
                     .unwrap_or_default()
-                    .to_string(),
-            ))
-        })
-        .unwrap_or_else(|| (raw_result.to_string(), String::new()));
-
-    let lines = stdout
-        .lines()
-        .chain(stderr.lines())
-        .map(str::trim_end)
-        .filter(|line| !line.trim().is_empty())
-        .map(str::to_string)
-        .collect::<Vec<_>>();
-    if lines.is_empty() {
-        return None;
-    }
-
-    let keep = 5;
-    let hidden_line_count = lines.len().saturating_sub(keep);
-    let visible = lines
-        .into_iter()
-        .skip(hidden_line_count)
-        .collect::<Vec<_>>();
-    Some(ShellOutputTail {
-        lines: visible,
-        hidden_line_count,
-    })
-}
-
-fn parse_tool_detail(
-    tool_name: &str,
-    arguments: Option<&serde_json::Value>,
-    cwd: Option<&str>,
-) -> ToolDetail {
-    let Some(args) = arguments else {
-        return ToolDetail::None;
-    };
-    // arguments can be a JSON string or an object
-    let obj = if let Some(s) = args.as_str() {
-        serde_json::from_str::<serde_json::Value>(s).unwrap_or_default()
-    } else {
-        args.clone()
-    };
-
-    let str_field = |key: &str| -> String {
-        obj.get(key)
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string()
-    };
-    let short = |path: &str| -> String {
-        let mut count = 0;
-        for (i, c) in path.char_indices().rev() {
-            if c == '/' {
-                count += 1;
-                if count == 2 {
-                    return path[i + 1..].to_string();
-                }
-            }
-        }
-        path.to_string()
-    };
-
-    match tool_name {
-        "edit" => {
-            let file = obj
-                .get("filePath")
-                .or_else(|| obj.get("file_path"))
-                .and_then(|v| v.as_str())
-                .unwrap_or("")
-                .to_string();
-            let old = obj
-                .get("oldString")
-                .or_else(|| obj.get("old_string"))
-                .and_then(|v| v.as_str())
-                .unwrap_or("")
-                .to_string();
-            let new = obj
-                .get("newString")
-                .or_else(|| obj.get("new_string"))
-                .and_then(|v| v.as_str())
-                .unwrap_or("")
-                .to_string();
-            let cached_lines = build_diff_lines(&old, &new, None);
-            ToolDetail::Edit {
-                file,
-                old,
-                new,
-                start_line: None,
-                cached_lines,
-            }
-        }
-        "multiedit" => {
-            let file = obj
-                .get("filePath")
-                .or_else(|| obj.get("file_path"))
-                .and_then(|v| v.as_str())
-                .unwrap_or("")
-                .to_string();
-            let sections = obj
-                .get("edits")
-                .and_then(|v| v.as_array())
-                .map(|edits| {
-                    edits
-                        .iter()
-                        .enumerate()
-                        .filter_map(|(idx, edit)| {
-                            let old = edit
-                                .get("oldString")
-                                .or_else(|| edit.get("old_string"))
-                                .and_then(|v| v.as_str())?;
-                            let new = edit
-                                .get("newString")
-                                .or_else(|| edit.get("new_string"))
-                                .and_then(|v| v.as_str())?;
-                            let replace_all = edit
-                                .get("replaceAll")
-                                .or_else(|| edit.get("replace_all"))
-                                .and_then(|v| v.as_bool())
-                                .unwrap_or(false);
-                            let suffix = if replace_all { " (all)" } else { "" };
-                            Some(DiffPreviewSection {
-                                header: format!("edit {}{}", idx + 1, suffix),
-                                old: old.to_string(),
-                                new: new.to_string(),
-                                start_line: None,
-                            })
-                        })
-                        .collect::<Vec<_>>()
-                })
-                .unwrap_or_default();
-            let edit_count = sections.len();
-            let cached_lines = build_sectioned_diff_lines(&sections, 6);
-            ToolDetail::MultiEdit {
-                file,
-                edit_count,
-                sections,
-                cached_lines,
-            }
-        }
-        "replace_symbol" => {
-            let replacements = obj
-                .get("replacements")
-                .and_then(|v| v.as_array())
-                .cloned()
-                .unwrap_or_default();
-            let root = obj.get("root").and_then(|v| v.as_str());
-            let title = replace_symbol_title(&replacements);
-            let sections = build_replace_symbol_sections(&replacements, root, cwd);
-            let cached_lines = build_sectioned_diff_lines(&sections, 4);
-            ToolDetail::ReplaceSymbol {
-                title,
-                sections,
-                cached_lines,
-            }
-        }
-        "write_file" => {
-            let path = str_field("path");
-            let content = str_field("content");
-            let cached_lines = build_write_lines(&content);
-            ToolDetail::WriteFile {
-                path,
-                content,
-                cached_lines,
-            }
-        }
-        "read_tool" => {
-            let path = str_field("path");
-            let offset = obj.get("offset").and_then(|v| v.as_u64()).unwrap_or(0);
-            let limit = obj
-                .get("limit")
-                .and_then(|v| v.as_u64())
-                .filter(|limit| *limit > 0)
-                .unwrap_or(DEFAULT_READ_TOOL_LIMIT);
-            let start_line = offset.saturating_add(1);
-            let end_line = offset.saturating_add(limit);
-            ToolDetail::ReadTool {
-                path,
-                start_line: Some(start_line),
-                end_line: Some(end_line),
-            }
-        }
-        "shell" => {
-            let command = shell_command_display(&obj);
-            let workdir = obj
-                .get("workdir")
-                .and_then(|v| v.as_str())
-                .map(str::to_string);
-            let cached_lines = build_shell_lines(&command, workdir.as_deref(), None);
-            ToolDetail::Shell {
-                command,
-                workdir,
-                output_tail: None,
-                cached_lines,
-            }
-        }
-        "search_text" => {
-            let pattern = str_field("pattern");
-            let path = str_field("path");
-            let include = str_field("include");
-            let location = if !include.is_empty() {
-                include
-            } else if !path.is_empty() {
-                short(&path).to_string()
             } else {
-                ".".into()
-            };
-            ToolDetail::Summary(format!("\"{}\" {}", pattern, location))
-        }
-        "glob" => {
-            let pattern = str_field("pattern");
-            let path = str_field("path");
-            if path.is_empty() {
-                ToolDetail::Summary(pattern)
-            } else {
-                ToolDetail::Summary(format!("{} in {}", pattern, short(&path)))
+                String::new()
             }
         }
-        "ls" => {
-            let path = str_field("path");
-            ToolDetail::Summary(if path.is_empty() {
-                ".".into()
-            } else {
-                short(&path).to_string()
+        serde_json::Value::Array(blocks) => blocks
+            .iter()
+            .filter_map(|block| {
+                (block.get("type").and_then(serde_json::Value::as_str) == Some("text"))
+                    .then(|| block.get("text").and_then(serde_json::Value::as_str))
+                    .flatten()
+                    .map(str::to_string)
             })
-        }
-        "delete_file" => {
-            let path = str_field("path");
-            ToolDetail::Summary(short(&path).to_string())
-        }
-        "browse" | "web_fetch" => {
-            let url = str_field("url");
-            ToolDetail::Summary(truncate_summary(&url, 60))
-        }
-        "apply_patch" => ToolDetail::Summary("patch".into()),
-        "delegate" => {
-            let agent = str_field("target_agent_id");
-            let objective = str_field("objective");
-            let obj_display = truncate_summary(&objective, 50);
-            let display = if agent.is_empty() {
-                obj_display
-            } else {
-                format!("({agent}) {obj_display}")
-            };
-            ToolDetail::Summary(display)
-        }
-        "language_query" => {
-            let action = str_field("action");
-            let uri = str_field("uri");
-            ToolDetail::Summary(format!("{} {}", action, short(&uri)))
-        }
-        "question" => ToolDetail::Summary("asking...".into()),
-        "todowrite" => {
-            if let Some(todos) = obj.get("todos").and_then(|v| v.as_array()) {
-                let lines: Vec<String> = todos
-                    .iter()
-                    .filter_map(|t| {
-                        let content = t.get("content").and_then(|v| v.as_str()).unwrap_or("");
-                        let status = t
-                            .get("status")
-                            .and_then(|v| v.as_str())
-                            .unwrap_or("pending");
-                        if content.is_empty() {
-                            return None;
-                        }
-                        let check = if status == "completed" { "x" } else { " " };
-                        Some(format!("[{check}] {content}"))
-                    })
-                    .collect();
-                if lines.is_empty() {
-                    ToolDetail::None
-                } else {
-                    ToolDetail::Summary(lines.join("\n"))
-                }
-            } else {
-                ToolDetail::None
-            }
-        }
-        "index" => {
-            let path = str_field("path");
-            ToolDetail::Summary(short(&path).to_string())
-        }
-        _ => ToolDetail::None,
+            .collect::<Vec<_>>()
+            .join(""),
+        _ => String::new(),
     }
 }
 
@@ -2871,345 +2328,6 @@ fn failed_tool_call_exists(
     })
 }
 
-fn reconcile_tool_call_start(
-    messages: &mut [ChatEntry],
-    tool_call_id: Option<&str>,
-    tool_name: &str,
-    start_detail: ToolDetail,
-) -> bool {
-    let Some(id) = tool_call_id else { return false };
-    let fallback_name = fallback_failed_tool_name(tool_name);
-    for entry in messages.iter_mut().rev() {
-        if let ChatEntry::ToolCall {
-            tool_call_id: Some(tid),
-            name,
-            detail,
-            ..
-        } = entry
-        {
-            if tid != id {
-                continue;
-            }
-            if name == &fallback_name {
-                *name = tool_name.to_string();
-            }
-            if matches!(detail, ToolDetail::None) && !matches!(start_detail, ToolDetail::None) {
-                *detail = start_detail;
-            }
-            return true;
-        }
-    }
-    false
-}
-
-fn mark_tool_call_failed(
-    messages: &mut [ChatEntry],
-    tool_call_id: Option<&str>,
-    tool_name: &str,
-) -> bool {
-    let Some(id) = tool_call_id else { return false };
-    for entry in messages.iter_mut().rev() {
-        if let ChatEntry::ToolCall {
-            tool_call_id: Some(tid),
-            name,
-            is_error,
-            ..
-        } = entry
-            && tid == id
-            && name == tool_name
-            && !*is_error
-        {
-            *is_error = true;
-            return true;
-        }
-    }
-    false
-}
-
-fn update_tool_detail(
-    messages: &mut [ChatEntry],
-    tool_call_id: Option<&str>,
-    result: &str,
-) -> bool {
-    let Some(id) = tool_call_id else { return false };
-    let parsed_result = serde_json::from_str::<serde_json::Value>(result).ok();
-
-    // walk backwards to find matching ToolCall
-    for entry in messages.iter_mut().rev() {
-        if let ChatEntry::ToolCall {
-            tool_call_id: Some(tid),
-            name,
-            detail,
-            ..
-        } = entry
-        {
-            if tid != id {
-                continue;
-            }
-            // edit tool: update line numbers from either legacy JSON or compact receipt output.
-            if let ToolDetail::Edit {
-                old,
-                new,
-                start_line,
-                cached_lines,
-                ..
-            } = detail
-            {
-                let compact_start =
-                    compact_receipt_old_starts(&tool_result_text(parsed_result.as_ref(), result))
-                        .into_iter()
-                        .next();
-                let json_start = parsed_result.as_ref().and_then(|obj| {
-                    obj.get("startLineOld")
-                        .and_then(|v| v.as_u64())
-                        .map(|n| n as usize)
-                });
-                if let Some(start) = json_start.or(compact_start) {
-                    *start_line = Some(start);
-                    *cached_lines = build_diff_lines(old, new, Some(start));
-                }
-            }
-            if let ToolDetail::MultiEdit {
-                sections,
-                cached_lines,
-                ..
-            } = detail
-            {
-                let starts =
-                    compact_receipt_old_starts(&tool_result_text(parsed_result.as_ref(), result));
-                if !starts.is_empty() {
-                    for (section, start) in sections.iter_mut().zip(starts) {
-                        section.start_line = Some(start);
-                    }
-                    *cached_lines = build_sectioned_diff_lines(sections, 6);
-                }
-            }
-            if name.starts_with("shell")
-                && let ToolDetail::Shell {
-                    command,
-                    workdir,
-                    output_tail,
-                    cached_lines,
-                } = detail
-                && let Some(tail) = shell_output_tail_from_result(parsed_result.as_ref(), result)
-            {
-                *output_tail = Some(tail);
-                *cached_lines =
-                    build_shell_lines(command, workdir.as_deref(), output_tail.as_ref());
-            }
-            if name == "read_tool"
-                && let ToolDetail::ReadTool {
-                    start_line,
-                    end_line,
-                    ..
-                } = detail
-                && let Some((first, last)) =
-                    read_tool_result_line_range(&tool_result_text(parsed_result.as_ref(), result))
-            {
-                *start_line = Some(first);
-                *end_line = Some(last);
-            }
-            // index tool: enrich with language and section counts
-            if name == "index" {
-                let text = parsed_result
-                    .as_ref()
-                    .map(content_to_string)
-                    .filter(|s| !s.is_empty())
-                    .unwrap_or_else(|| result.to_string());
-                if let Some(summary_parts) = index_outline_summary_parts(&text)
-                    && let ToolDetail::Summary(path) = detail
-                {
-                    let suffix = format!(" ({})", summary_parts.join(", "));
-                    if !path.ends_with(&suffix) {
-                        path.push_str(&suffix);
-                    }
-                }
-            }
-            // search_text tool: append backend result counts from the compact footer.
-            if name == "search_text" {
-                let text = parsed_result
-                    .as_ref()
-                    .map(content_to_string)
-                    .filter(|s| !s.is_empty())
-                    .unwrap_or_else(|| result.to_string());
-                if let Some(footer) = search_text_footer(&text)
-                    && let ToolDetail::Summary(summary) = detail
-                {
-                    let suffix = format!(" {footer}");
-                    if !summary.ends_with(&suffix) {
-                        summary.push_str(&suffix);
-                    }
-                }
-            }
-            return true;
-        }
-    }
-    false
-}
-
-fn tool_result_text(parsed_result: Option<&serde_json::Value>, raw_result: &str) -> String {
-    parsed_result
-        .map(content_to_string)
-        .filter(|s| !s.is_empty())
-        .unwrap_or_else(|| raw_result.to_string())
-}
-
-fn read_tool_result_line_range(text: &str) -> Option<(u64, u64)> {
-    let mut first = None;
-    let mut last = None;
-
-    for line in text.lines() {
-        let Some((prefix, _)) = line.split_once('|') else {
-            continue;
-        };
-        if prefix.len() != 5 || !prefix.bytes().all(|b| b.is_ascii_digit()) {
-            continue;
-        }
-        let Ok(line_number) = prefix.parse::<u64>() else {
-            continue;
-        };
-        first.get_or_insert(line_number);
-        last = Some(line_number);
-    }
-
-    first.zip(last)
-}
-
-fn compact_receipt_old_starts(text: &str) -> Vec<usize> {
-    text.lines()
-        .filter_map(|line| {
-            let line = line.trim();
-            if !line.starts_with("H ") {
-                return None;
-            }
-            let old_part = line
-                .split_whitespace()
-                .find(|part| part.starts_with("old="))?;
-            old_part
-                .strip_prefix("old=")?
-                .split(',')
-                .next()?
-                .parse::<usize>()
-                .ok()
-        })
-        .collect()
-}
-
-fn search_text_footer(text: &str) -> Option<&str> {
-    let footer = text
-        .lines()
-        .rev()
-        .find(|line| !line.trim().is_empty())?
-        .trim();
-    if !footer.starts_with('(') || !footer.ends_with(')') {
-        return None;
-    }
-
-    let inner = &footer[1..footer.len() - 1];
-    let parts: Vec<&str> = inner.split(", ").collect();
-    if !(parts.len() == 2 || parts.len() == 3) {
-        return None;
-    }
-    if parts.len() == 3 && parts[2] != "truncated" {
-        return None;
-    }
-
-    let file_label = if parse_counted_label(parts[0], "file")? == 1 {
-        "file"
-    } else {
-        "files"
-    };
-    let match_label = if parse_counted_label(parts[1], "match")? == 1 {
-        "match"
-    } else {
-        "matches"
-    };
-    if !parts[0].ends_with(file_label) || !parts[1].ends_with(match_label) {
-        return None;
-    }
-
-    Some(footer)
-}
-
-fn parse_counted_label(text: &str, singular: &str) -> Option<usize> {
-    let (count, label) = text.split_once(' ')?;
-    let count = count.parse::<usize>().ok()?;
-    let plural = if singular.ends_with("ch") {
-        format!("{singular}es")
-    } else {
-        format!("{singular}s")
-    };
-    (label == singular || label == plural).then_some(count)
-}
-
-fn index_outline_summary_parts(text: &str) -> Option<Vec<String>> {
-    let lang = text
-        .lines()
-        .find(|l| l.starts_with("language:"))
-        .and_then(|l| l.split_once(':'))
-        .map(|(_, value)| value.trim())
-        .filter(|value| !value.is_empty());
-    let mut counts: Vec<(String, usize)> = Vec::new();
-    let mut current_section: Option<String> = None;
-    let mut current_count = 0usize;
-    for line in text.lines() {
-        if line.ends_with(':') && !line.starts_with(' ') && !line.starts_with('\t') {
-            if let Some(ref section) = current_section
-                && current_count > 0
-            {
-                counts.push((section.clone(), current_count));
-            }
-            let section = line.trim_end_matches(':');
-            current_section =
-                (!matches!(section, "path" | "language")).then(|| section.to_string());
-            current_count = 0;
-        } else if line.starts_with("  - ") && current_section.is_some() {
-            current_count += 1;
-        }
-    }
-    if let Some(ref section) = current_section
-        && current_count > 0
-    {
-        counts.push((section.clone(), current_count));
-    }
-    let mut summary_parts: Vec<String> = Vec::new();
-    if let Some(l) = lang {
-        summary_parts.push(l.to_string());
-    }
-    for (sec, n) in counts {
-        summary_parts.push(format!("{n} {sec}"));
-    }
-    (!summary_parts.is_empty()).then_some(summary_parts)
-}
-
-fn content_to_string(v: &serde_json::Value) -> String {
-    match v {
-        serde_json::Value::String(s) => s.clone(),
-        serde_json::Value::Object(obj) => {
-            if obj.get("type").and_then(|t| t.as_str()) == Some("text") {
-                obj.get("text")
-                    .and_then(|t| t.as_str())
-                    .map(String::from)
-                    .unwrap_or_default()
-            } else {
-                String::new()
-            }
-        }
-        serde_json::Value::Array(arr) => arr
-            .iter()
-            .filter_map(|block| {
-                if block.get("type").and_then(|t| t.as_str()) == Some("text") {
-                    block.get("text").and_then(|t| t.as_str()).map(String::from)
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<_>>()
-            .join(""),
-        _ => String::new(),
-    }
-}
-
 // ── scroll_tests ─────────────────────────────────────────────────────────────
 
 #[cfg(test)]
@@ -3262,821 +2380,6 @@ mod fork_result_tests {
         assert_eq!(app.pending_fork_message_id, None);
         assert_eq!(app.popup, Popup::ForkTurnSelect);
         assert!(app.status.contains("nope"));
-    }
-}
-
-#[cfg(test)]
-mod tool_detail_tests {
-    use super::{
-        failed_tool_call_exists, mark_tool_call_failed, parse_tool_detail, push_error_message,
-        reconcile_tool_call_start, update_tool_detail,
-    };
-    use crate::app::{App, ChatEntry, ToolDetail};
-    use crate::protocol::{EventKind, RawServerMsg};
-
-    fn summary_text(entry: &ChatEntry) -> &str {
-        match entry {
-            ChatEntry::ToolCall {
-                detail: ToolDetail::Summary(s),
-                ..
-            } => s,
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn duplicate_error_replay_does_not_append_stale_card() {
-        let mut messages = Vec::new();
-
-        assert!(push_error_message(&mut messages, "LLM streaming error"));
-        messages.push(ChatEntry::User {
-            text: "continue".into(),
-            message_id: Some("user-continue".into()),
-        });
-        messages.push(ChatEntry::Assistant {
-            content: "latest assistant summary".into(),
-            thinking: None,
-            message_id: Some("assistant-latest".into()),
-        });
-
-        assert!(!push_error_message(&mut messages, "LLM streaming error"));
-        assert_eq!(messages.len(), 3, "replayed error must not append");
-        assert_eq!(
-            messages
-                .iter()
-                .filter(
-                    |entry| matches!(entry, ChatEntry::Error(text) if text == "LLM streaming error")
-                )
-                .count(),
-            1
-        );
-        assert!(matches!(
-            messages.last(),
-            Some(ChatEntry::Assistant { content, .. }) if content == "latest assistant summary"
-        ));
-    }
-
-    #[test]
-    fn distinct_error_messages_are_preserved() {
-        let mut messages = Vec::new();
-
-        assert!(push_error_message(&mut messages, "first error"));
-        assert!(push_error_message(&mut messages, "second error"));
-
-        assert_eq!(messages.len(), 2);
-        assert!(matches!(&messages[0], ChatEntry::Error(text) if text == "first error"));
-        assert!(matches!(&messages[1], ChatEntry::Error(text) if text == "second error"));
-    }
-
-    #[test]
-    fn event_error_replay_does_not_append_stale_card() {
-        let mut app = App::new();
-        let error = EventKind::Error {
-            message: "LLM streaming error".into(),
-        };
-
-        app.handle_event_kind(&error, false, None);
-        app.messages.push(ChatEntry::User {
-            text: "continue".into(),
-            message_id: Some("user-continue".into()),
-        });
-        app.messages.push(ChatEntry::Assistant {
-            content: "latest assistant summary".into(),
-            thinking: None,
-            message_id: Some("assistant-latest".into()),
-        });
-        app.handle_event_kind(&error, true, None);
-
-        assert_eq!(app.messages.len(), 3, "replayed error must not append");
-        assert_eq!(
-            app.messages
-                .iter()
-                .filter(
-                    |entry| matches!(entry, ChatEntry::Error(text) if text == "LLM streaming error")
-                )
-                .count(),
-            1
-        );
-        assert!(matches!(
-            app.messages.last(),
-            Some(ChatEntry::Assistant { content, .. }) if content == "latest assistant summary"
-        ));
-    }
-
-    #[test]
-    fn raw_server_error_duplicate_does_not_append() {
-        let mut app = App::new();
-        let raw = || RawServerMsg {
-            msg_type: "error".into(),
-            data: Some(serde_json::json!({ "message": "raw server error" })),
-        };
-
-        app.handle_server_msg(raw());
-        app.messages.push(ChatEntry::Assistant {
-            content: "still latest".into(),
-            thinking: None,
-            message_id: None,
-        });
-        app.handle_server_msg(raw());
-
-        assert_eq!(app.messages.len(), 2, "duplicate raw error must not append");
-        assert!(matches!(
-            app.messages.last(),
-            Some(ChatEntry::Assistant { content, .. }) if content == "still latest"
-        ));
-    }
-
-    #[test]
-    fn delegate_tool_shows_agent_and_objective() {
-        let args = serde_json::json!({
-            "target_agent_id": "coder",
-            "objective": "List the contents of /tmp"
-        });
-        let detail = parse_tool_detail("delegate", Some(&args), None);
-        match detail {
-            ToolDetail::Summary(s) => {
-                assert!(s.contains("coder"), "must contain agent name, got: {s}");
-                assert!(
-                    s.contains("List the contents"),
-                    "must contain objective, got: {s}"
-                );
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn delegate_tool_without_agent_shows_objective_only() {
-        let args = serde_json::json!({
-            "objective": "Do something"
-        });
-        let detail = parse_tool_detail("delegate", Some(&args), None);
-        match detail {
-            ToolDetail::Summary(s) => {
-                assert!(
-                    s.contains("Do something"),
-                    "must contain objective, got: {s}"
-                );
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    fn assert_summary_truncates_with_ellipsis(tool_name: &str, args: serde_json::Value) {
-        let detail = parse_tool_detail(tool_name, Some(&args), None);
-
-        match detail {
-            ToolDetail::Summary(s) => {
-                assert!(s.ends_with(crate::ui::ELLIPSIS), "got: {s}");
-                assert!(s.is_char_boundary(s.len()), "got: {s}");
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn shell_tool_preserves_full_utf8_command() {
-        let command =
-            "cat > check/kimi.md << 'EOF'\n# Review: feat/profiles\n\n## 🔴 Critical / High";
-        let args = serde_json::json!({ "command": command });
-        let detail = parse_tool_detail("shell", Some(&args), None);
-
-        match detail {
-            ToolDetail::Shell { command: got, .. } => assert_eq!(got, command),
-            other => panic!("expected Shell, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn shell_tool_uses_only_explicit_workdir() {
-        let explicit = parse_tool_detail(
-            "shell",
-            Some(&serde_json::json!({ "command": "pwd", "workdir": "/workspace/project" })),
-            Some("/session/cwd"),
-        );
-        match explicit {
-            ToolDetail::Shell { workdir, .. } => {
-                assert_eq!(workdir.as_deref(), Some("/workspace/project"))
-            }
-            other => panic!("expected Shell, got: {other:?}"),
-        }
-
-        let implicit = parse_tool_detail(
-            "shell",
-            Some(&serde_json::json!({ "command": "pwd" })),
-            Some("/session/cwd"),
-        );
-        match implicit {
-            ToolDetail::Shell { workdir, .. } => assert_eq!(workdir, None),
-            other => panic!("expected Shell, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn read_tool_uses_one_based_inclusive_display_range() {
-        let detail = parse_tool_detail(
-            "read_tool",
-            Some(&serde_json::json!({
-                "path": "src/server_msg.rs",
-                "offset": 2134,
-                "limit": 71,
-            })),
-            None,
-        );
-
-        match detail {
-            ToolDetail::ReadTool {
-                path,
-                start_line,
-                end_line,
-            } => {
-                assert_eq!(path, "src/server_msg.rs");
-                assert_eq!(start_line, Some(2135));
-                assert_eq!(end_line, Some(2205));
-            }
-            other => panic!("expected ReadTool, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn read_tool_defaults_missing_offset_and_limit_for_display_range() {
-        let detail = parse_tool_detail(
-            "read_tool",
-            Some(&serde_json::json!({
-                "path": "apps/portal/priv/static/assets/css/app.css",
-                "limit": 300,
-            })),
-            None,
-        );
-
-        match detail {
-            ToolDetail::ReadTool {
-                start_line,
-                end_line,
-                ..
-            } => {
-                assert_eq!(start_line, Some(1));
-                assert_eq!(end_line, Some(300));
-            }
-            other => panic!("expected ReadTool, got: {other:?}"),
-        }
-
-        let default_limit = parse_tool_detail(
-            "read_tool",
-            Some(&serde_json::json!({ "path": "README.md" })),
-            None,
-        );
-        match default_limit {
-            ToolDetail::ReadTool {
-                start_line,
-                end_line,
-                ..
-            } => {
-                assert_eq!(start_line, Some(1));
-                assert_eq!(end_line, Some(super::DEFAULT_READ_TOOL_LIMIT));
-            }
-            other => panic!("expected ReadTool, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn read_tool_parses_json_string_arguments_for_display_range() {
-        let args = serde_json::Value::String(
-            r#"{"path":"apps/admin/lib/admin_web/components/layouts/root.html.heex","root":".","limit":220}"#
-                .into(),
-        );
-        let detail = parse_tool_detail("read_tool", Some(&args), None);
-
-        match detail {
-            ToolDetail::ReadTool {
-                path,
-                start_line,
-                end_line,
-            } => {
-                assert_eq!(
-                    path,
-                    "apps/admin/lib/admin_web/components/layouts/root.html.heex"
-                );
-                assert_eq!(start_line, Some(1));
-                assert_eq!(end_line, Some(220));
-            }
-            other => panic!("expected ReadTool, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn session_loaded_replay_audit_preserves_full_utf8_shell_command() {
-        let command =
-            "cat > check/kimi.md << 'EOF'\n# Review: feat/profiles\n\n## 🔴 Critical / High";
-        let mut app = App::new();
-
-        app.handle_server_msg(RawServerMsg {
-            msg_type: "session_loaded".into(),
-            data: Some(serde_json::json!({
-                "session_id": "sess-utf8",
-                "agent_id": "coder",
-                "audit": {
-                    "events": [{
-                        "kind": { "type": "tool_call_start", "data": {
-                            "tool_call_id": "tool-utf8",
-                            "tool_name": "shell",
-                            "arguments": { "command": command }
-                        }},
-                        "timestamp": null
-                    }]
-                },
-                "undo_stack": []
-            })),
-        });
-
-        assert_eq!(app.messages.len(), 1);
-        match &app.messages[0] {
-            ChatEntry::ToolCall {
-                detail: ToolDetail::Shell { command: got, .. },
-                ..
-            } => assert_eq!(got, command),
-            other => panic!("expected Shell tool call, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn shell_tool_result_keeps_last_five_output_lines() {
-        let detail = parse_tool_detail(
-            "shell",
-            Some(&serde_json::json!({ "command": "cargo test" })),
-            None,
-        );
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("shell-tail".into()),
-            name: "shell".into(),
-            is_error: false,
-            detail,
-        }];
-        let result = serde_json::json!({
-            "exit_code": 0,
-            "stdout": "out1\nout2\nout3\nout4\nout5\nout6\n",
-            "stderr": "err1\nerr2\n",
-        })
-        .to_string();
-
-        assert!(update_tool_detail(
-            &mut messages,
-            Some("shell-tail"),
-            &result
-        ));
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail:
-                    ToolDetail::Shell {
-                        output_tail: Some(tail),
-                        cached_lines,
-                        ..
-                    },
-                ..
-            } => {
-                assert_eq!(tail.hidden_line_count, 3);
-                assert_eq!(tail.lines, ["out4", "out5", "out6", "err1", "err2"]);
-                let rendered = cached_lines
-                    .iter()
-                    .flat_map(|line| line.spans.iter().map(|span| span.content.as_ref()))
-                    .collect::<String>();
-                assert!(rendered.contains("output tail:"));
-                assert!(rendered.contains("err2"));
-            }
-            other => panic!("expected shell tail, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn read_tool_result_refines_range_to_actual_line_numbers() {
-        let detail = parse_tool_detail(
-            "read_tool",
-            Some(&serde_json::json!({
-                "path": "apps/portal/lib/portal_web/components/layouts/root.html.heex",
-                "limit": 260,
-            })),
-            None,
-        );
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("read-lines".into()),
-            name: "read_tool".into(),
-            is_error: false,
-            detail,
-        }];
-        let result = "<path>/projects/querymt-org/service/apps/portal/lib/portal_web/components/layouts/root.html.heex</path>\n<type>file</type>\n<content>\n00001| <!DOCTYPE html>\n00002| <html lang=\"en\">\n00027| </html>\n\n(End of file - total 27 lines)\n</content>";
-
-        assert!(update_tool_detail(
-            &mut messages,
-            Some("read-lines"),
-            result
-        ));
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail:
-                    ToolDetail::ReadTool {
-                        start_line,
-                        end_line,
-                        ..
-                    },
-                ..
-            } => {
-                assert_eq!(*start_line, Some(1));
-                assert_eq!(*end_line, Some(27));
-            }
-            other => panic!("expected read_tool detail, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn read_tool_result_refines_range_from_json_text_content() {
-        let detail = parse_tool_detail(
-            "read_tool",
-            Some(&serde_json::json!({
-                "path": "apps/portal/lib/portal_web/components/core_components.ex",
-                "offset": 40,
-                "limit": 390,
-            })),
-            None,
-        );
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("read-json-lines".into()),
-            name: "read_tool".into(),
-            is_error: false,
-            detail,
-        }];
-        let result = serde_json::json!([{
-            "type": "text",
-            "text": "<path>/repo/core_components.ex</path>\n<type>file</type>\n<content>\n00041|   attr :id, :string\n00042|   slot :inner_block\n00430| end\n\n(File has more lines. Use 'offset' parameter to read beyond line 430)\n</content>"
-        }])
-        .to_string();
-
-        assert!(update_tool_detail(
-            &mut messages,
-            Some("read-json-lines"),
-            &result
-        ));
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail:
-                    ToolDetail::ReadTool {
-                        start_line,
-                        end_line,
-                        ..
-                    },
-                ..
-            } => {
-                assert_eq!(*start_line, Some(41));
-                assert_eq!(*end_line, Some(430));
-            }
-            other => panic!("expected read_tool detail, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn multiedit_tool_result_updates_preview_line_numbers() {
-        let detail = parse_tool_detail(
-            "multiedit",
-            Some(&serde_json::json!({
-                "filePath": "src/lib.rs",
-                "edits": [
-                    { "oldString": "aaa\n", "newString": "bbb\n" },
-                    { "oldString": "ccc\n", "newString": "ddd\n" }
-                ]
-            })),
-            None,
-        );
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("multi-lines".into()),
-            name: "multiedit".into(),
-            is_error: false,
-            detail,
-        }];
-        let result = "OK paths=1 edits=2 added=2 deleted=2\nP src/lib.rs\nH replace old=10,1 new=10,1 +1 -1\nH replace old=20,1 new=20,1 +1 -1";
-
-        assert!(update_tool_detail(
-            &mut messages,
-            Some("multi-lines"),
-            result
-        ));
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail: ToolDetail::MultiEdit { sections, .. },
-                ..
-            } => {
-                assert_eq!(sections[0].start_line, Some(10));
-                assert_eq!(sections[1].start_line, Some(20));
-            }
-            other => panic!("expected multiedit detail, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn browse_tool_truncates_utf8_url_on_char_boundary() {
-        let url = "https://example.com/docs/aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa🔴/page";
-        let args = serde_json::json!({ "url": url });
-
-        assert_summary_truncates_with_ellipsis("browse", args);
-    }
-
-    #[test]
-    fn delegate_tool_truncates_utf8_objective_on_char_boundary() {
-        let objective = "Review this objective with multibyte marker aaaaa🔴 done";
-        assert!(!objective.is_char_boundary(50));
-        let args = serde_json::json!({ "objective": objective });
-
-        assert_summary_truncates_with_ellipsis("delegate", args);
-    }
-
-    #[test]
-    fn index_tool_shows_short_path_from_arguments() {
-        let args = serde_json::json!({"path": "/home/user/project/src/main.rs"});
-        let detail = parse_tool_detail("index", Some(&args), None);
-        match detail {
-            ToolDetail::Summary(s) => {
-                assert_eq!(s, "src/main.rs");
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn failed_tool_end_marks_existing_tool_in_place() {
-        let mut messages = vec![
-            ChatEntry::ToolCall {
-                tool_call_id: Some("tool-1".into()),
-                name: "shell".into(),
-                is_error: false,
-                detail: ToolDetail::Summary("echo ok".into()),
-            },
-            ChatEntry::Assistant {
-                content: "done".into(),
-                thinking: None,
-                message_id: None,
-            },
-        ];
-
-        assert!(mark_tool_call_failed(
-            &mut messages,
-            Some("tool-1"),
-            "shell"
-        ));
-
-        assert_eq!(messages.len(), 2, "must not append a stale failed badge");
-        match &messages[0] {
-            ChatEntry::ToolCall { name, is_error, .. } => {
-                assert_eq!(name, "shell");
-                assert!(*is_error);
-            }
-            other => panic!("expected ToolCall, got: {other:?}"),
-        }
-        assert!(matches!(messages[1], ChatEntry::Assistant { .. }));
-    }
-
-    #[test]
-    fn failed_tool_fallback_duplicate_check_is_idempotent() {
-        let messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("missing-start".into()),
-            name: "ls (failed)".into(),
-            is_error: true,
-            detail: ToolDetail::None,
-        }];
-
-        assert!(!mark_tool_call_failed(
-            &mut messages.clone(),
-            Some("missing-start"),
-            "ls"
-        ));
-        assert!(failed_tool_call_exists(
-            &messages,
-            Some("missing-start"),
-            "ls"
-        ));
-    }
-
-    #[test]
-    fn tool_start_reconciles_failed_fallback_in_place() {
-        let mut messages = vec![
-            ChatEntry::Assistant {
-                content: "before".into(),
-                thinking: None,
-                message_id: None,
-            },
-            ChatEntry::ToolCall {
-                tool_call_id: Some("missing-start".into()),
-                name: "shell (failed)".into(),
-                is_error: true,
-                detail: ToolDetail::None,
-            },
-        ];
-        let detail = parse_tool_detail(
-            "shell",
-            Some(&serde_json::json!({
-                "command": "cargo test tool_detail_tests"
-            })),
-            None,
-        );
-
-        assert!(reconcile_tool_call_start(
-            &mut messages,
-            Some("missing-start"),
-            "shell",
-            detail
-        ));
-        assert_eq!(
-            messages.len(),
-            2,
-            "start must not append a second tool entry"
-        );
-        match &messages[1] {
-            ChatEntry::ToolCall {
-                name,
-                is_error,
-                detail,
-                ..
-            } => {
-                assert_eq!(name, "shell");
-                assert!(*is_error);
-                assert!(
-                    matches!(detail, ToolDetail::Shell { command, .. } if command == "cargo test tool_detail_tests")
-                );
-            }
-            other => panic!("expected ToolCall, got: {other:?}"),
-        }
-        assert!(failed_tool_call_exists(
-            &messages,
-            Some("missing-start"),
-            "shell"
-        ));
-    }
-
-    #[test]
-    fn index_tool_enriched_with_language_and_counts() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("tc-1".into()),
-            name: "index".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("src/main.rs".into()),
-        }];
-        let result = "path: /home/user/project/src/main.rs\nlanguage: rust\n\nimports:\n  - use std::io [1]\n  - use std::fs [2]\ntypes:\n  - struct Foo [10-20]\n    - field: i32 [11]\nfunctions:\n  - fn main [30-40]\n  - fn helper [50-60]\n";
-        update_tool_detail(&mut messages, Some("tc-1"), result);
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail: ToolDetail::Summary(s),
-                ..
-            } => {
-                assert_eq!(
-                    s, "src/main.rs (rust, 2 imports, 1 types, 2 functions)",
-                    "must summarize raw outline result"
-                );
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn index_tool_enrichment_is_idempotent() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("tc-dup".into()),
-            name: "index".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("src/main.rs".into()),
-        }];
-        let result = "path: /home/user/project/src/main.rs\nlanguage: rust\n\nimports:\n  - use std::io [1]\ntypes:\n  - struct Foo [10-20]\nfunctions:\n  - fn main [30-40]\n";
-        update_tool_detail(&mut messages, Some("tc-dup"), result);
-        update_tool_detail(&mut messages, Some("tc-dup"), result);
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail: ToolDetail::Summary(s),
-                ..
-            } => {
-                assert_eq!(
-                    s, "src/main.rs (rust, 1 imports, 1 types, 1 functions)",
-                    "must not append duplicate index summary suffixes"
-                );
-                assert_eq!(
-                    s.matches("(rust, 1 imports, 1 types, 1 functions)").count(),
-                    1
-                );
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn index_tool_handles_json_text_blocks_and_skips_empty_sections() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("tc-2".into()),
-            name: "index".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("src/lib.rs".into()),
-        }];
-        let result = r#"[{"type":"text","text":"path: /home/user/project/src/lib.rs\nlanguage: python\n\nimports:\n  - os [1]\ntypes:\nfunctions:\n  - fn foo [10-20]\n"}]"#;
-        update_tool_detail(&mut messages, Some("tc-2"), result);
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail: ToolDetail::Summary(s),
-                ..
-            } => {
-                assert_eq!(
-                    s, "src/lib.rs (python, 1 imports, 1 functions)",
-                    "must summarize JSON text blocks and skip empty sections"
-                );
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn index_tool_graceful_on_malformed_result() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("tc-3".into()),
-            name: "index".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("src/lib.rs".into()),
-        }];
-        let result = "totally wrong";
-        update_tool_detail(&mut messages, Some("tc-3"), result);
-        match &messages[0] {
-            ChatEntry::ToolCall {
-                detail: ToolDetail::Summary(s),
-                ..
-            } => {
-                assert_eq!(
-                    s, "src/lib.rs",
-                    "must preserve original path on malformed result"
-                );
-            }
-            other => panic!("expected Summary, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn search_text_tool_enriched_with_raw_backend_footer() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("search-1".into()),
-            name: "search_text".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("\"needle\" *.rs".into()),
-        }];
-        let result = "config.rs\n13:needle\n\nmain.rs\n7:needle\n(5 files, 28 matches)";
-
-        update_tool_detail(&mut messages, Some("search-1"), result);
-
-        assert_eq!(
-            summary_text(&messages[0]),
-            "\"needle\" *.rs (5 files, 28 matches)"
-        );
-    }
-
-    #[test]
-    fn search_text_tool_enriched_with_truncated_footer() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("search-truncated".into()),
-            name: "search_text".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("\"needle\" src".into()),
-        }];
-        let result = "main.rs\n7:needle\n(1 file, 100 matches, truncated)";
-
-        update_tool_detail(&mut messages, Some("search-truncated"), result);
-
-        assert_eq!(
-            summary_text(&messages[0]),
-            "\"needle\" src (1 file, 100 matches, truncated)"
-        );
-    }
-
-    #[test]
-    fn search_text_tool_enrichment_is_idempotent() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("search-dup".into()),
-            name: "search_text".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("\"needle\" .".into()),
-        }];
-        let result = "(0 files, 0 matches)";
-
-        update_tool_detail(&mut messages, Some("search-dup"), result);
-        update_tool_detail(&mut messages, Some("search-dup"), result);
-
-        let summary = summary_text(&messages[0]);
-        assert_eq!(summary, "\"needle\" . (0 files, 0 matches)");
-        assert_eq!(summary.matches("(0 files, 0 matches)").count(), 1);
-    }
-
-    #[test]
-    fn search_text_tool_enriched_from_json_text_block_footer() {
-        let mut messages = vec![ChatEntry::ToolCall {
-            tool_call_id: Some("search-json".into()),
-            name: "search_text".into(),
-            is_error: false,
-            detail: ToolDetail::Summary("\"needle\" .".into()),
-        }];
-        let result = r#"[{"type":"text","text":"main.rs\n7:needle\n(1 file, 1 match)"}]"#;
-
-        update_tool_detail(&mut messages, Some("search-json"), result);
-
-        assert_eq!(summary_text(&messages[0]), "\"needle\" . (1 file, 1 match)");
     }
 }
